integrations/deli/requirements.txt
boto3==1.28.57
cidr-trie==3.1.2
cryptocode==0.1
dpath==2.1.6
facebook-business==15.0.0
flask==3.0.0
fuzzywuzzy==0.18.0
glob2==0.7
google-ads==26.0.1
idna==3.4
kafka-python==2.0.2
mysql-connector-python==9.2.0
pandas==1.3.5
protobuf==5.26.1
pygtail==0.14.0
scipy==1.7.3
sklearn==0.0
requests==2.31.0
ua-parser==0.18.0
user-agents==2.2.0
uwsgi==2.0.22
six==1.16.0
httpagentparser==1.9.5
